second set of changes to improve Sonar analysis
