+++ operator/src/main/java/oracle/kubernetes/weblogic/domain/model/KubernetesListObjectImpl.java
@@ -1,0 +1,101 @@
+// Copyright (c) 2022, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package oracle.kubernetes.weblogic.domain.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+import io.kubernetes.client.common.KubernetesListObject;
+import io.kubernetes.client.openapi.models.V1ListMeta;
+import jakarta.validation.Valid;
+
+public abstract class KubernetesListObjectImpl implements KubernetesListObject {
+  /**
+   * APIVersion defines the versioned schema of this representation of an object. Servers should
+   * convert recognized schemas to the latest internal value, and may reject unrecognized values.
+   * More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources.
+   */
+  @SerializedName("apiVersion")
+  @Expose
+  protected String apiVersion;
+  /**
+   * Kind is a string value representing the REST resource this object represents. Servers may infer
+   * this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More
+   * info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds.
+   */
+  @SerializedName("kind")
+  @Expose
+  protected String kind;
+  /**
+   * Standard list metadata. More info:
+   * https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds.
+   */
+  @SerializedName("metadata")
+  @Expose
+  @Valid
+  protected V1ListMeta metadata;
+
+  /**
+   * APIVersion defines the versioned schema of this representation of an object. Servers should
+   * convert recognized schemas to the latest internal value, and may reject unrecognized values.
+   * More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources.
+   *
+   * @return API version
+   */
+  public String getApiVersion() {
+    return apiVersion;
+  }
+
+  /**
+   * APIVersion defines the versioned schema of this representation of an object. Servers should
+   * convert recognized schemas to the latest internal value, and may reject unrecognized values.
+   * More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources.
+   *
+   * @param apiVersion API version
+   */
+  public void setApiVersion(String apiVersion) {
+    this.apiVersion = apiVersion;
+  }
+
+  /**
+   * Kind is a string value representing the REST resource this object represents. Servers may infer
+   * this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More
+   * info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds
+   *
+   * @return kind
+   */
+  public String getKind() {
+    return kind;
+  }
+
+  /**
+   * Kind is a string value representing the REST resource this object represents. Servers may infer
+   * this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More
+   * info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds
+   *
+   * @param kind kind
+   */
+  public void setKind(String kind) {
+    this.kind = kind;
+  }
+
+  /**
+   * Standard list metadata. More info:
+   * https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds
+   *
+   * @return metadata
+   */
+  public V1ListMeta getMetadata() {
+    return metadata;
+  }
+
+  /**
+   * Standard list metadata. More info:
+   * https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds
+   *
+   * @param metadata metadata
+   */
+  public void setMetadata(V1ListMeta metadata) {
+    this.metadata = metadata;
+  }
+}
